Move deferred examples to Librum::Components::RSpec.
- Add Plumbum::StubProvider to test configuration.
- Update RSpec::SleepingKingStudios dependency.
