+++ lib/librum/components/rspec/deferred.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require 'librum/components/rspec'
+
+module Librum::Components::RSpec
+  # Namespace for deferred example groups.
+  module Deferred; end
+end
