+++ lib/librum/components/rspec/deferred/options_examples.rb
@@ -1,8 +1,9 @@
 # frozen_string_literal: true
 
-require 'support/deferred'
+require 'librum/components/rspec/deferred'
 
-module Spec::Support::Deferred
+module Librum::Components::RSpec::Deferred
+  # Deferred examples verifying component options.
   module OptionsExamples
     include RSpec::SleepingKingStudios::Deferred::Provider
 
@@ -14,6 +15,31 @@ module Spec::Support::Deferred
       end
     end
 
+    deferred_examples 'should check for duplicate options' do
+      include RSpec::SleepingKingStudios::Deferred::Dependencies
+
+      depends_on :name,         'the name of the defined option'
+      depends_on :meta_options, 'the options passed to .option()'
+
+      wrap_deferred 'when the option is defined' do
+        let(:error_message) do
+          option_name = "#{name}#{'?' if meta_options[:boolean]}"
+
+          "unable to define option ##{option_name} - the option is already " \
+            "defined on #{described_class.name}"
+        end
+
+        it 'should raise an exception' do
+          expect { described_class.option(name, **meta_options) }
+            .to raise_error(
+              Librum::Components::Options::DuplicateOptionError,
+              error_message
+            )
+        end
+      end
+    end
+
+    # Deferred examples that validate an existing option for a component.
     deferred_examples 'should define component option' \
     do |name, boolean: false, default: nil, value: 'value'|
       name = name.to_s
@@ -54,6 +80,104 @@ module Spec::Support::Deferred
           let(:component_options) { super().merge(name.intern => value) }
 
           it { expect(component.send(name)).to be == value }
+        end
+      end
+    end
+
+    # Deferred examples that validate defining a new option.
+    deferred_examples 'should define the configured option' do
+      include RSpec::SleepingKingStudios::Deferred::Dependencies
+
+      depends_on :name,         'the name of the defined option'
+      depends_on :meta_options, 'the options passed to .option()'
+
+      describe 'with boolean: true' do
+        let(:meta_options) { super().merge(boolean: true) }
+
+        it 'should return the name of the generated method' do
+          expect(described_class.option(name, **meta_options))
+            .to be :"#{name}?"
+        end
+
+        include_deferred 'should check for duplicate options'
+
+        wrap_deferred 'when the option is defined' do
+          include_deferred 'should define component option',
+            'example_option?',
+            boolean: true,
+            default: false
+        end
+
+        describe 'with default: a Proc' do
+          let(:meta_options)     { super().merge(default: -> { 'value' }) }
+          let(:expected_default) { 'value' }
+
+          wrap_deferred 'when the option is defined' do
+            include_deferred 'should define component option',
+              'example_option',
+              boolean: true,
+              default: -> { expected_default }
+          end
+        end
+
+        describe 'with default: value' do
+          let(:meta_options)     { super().merge(default: 'value') }
+          let(:expected_default) { 'value' }
+
+          wrap_deferred 'when the option is defined' do
+            include_deferred 'should define component option',
+              'example_option',
+              boolean: true,
+              default: 'value'
+          end
+        end
+      end
+
+      describe 'with default: a Proc' do
+        let(:meta_options)     { super().merge(default: -> { 'value' }) }
+        let(:expected_default) { 'value' }
+
+        wrap_deferred 'when the option is defined' do
+          include_deferred 'should define component option',
+            'example_option',
+            default: -> { expected_default }
+        end
+      end
+
+      describe 'with default: value' do
+        let(:meta_options)     { super().merge(default: 'value') }
+        let(:expected_default) { 'value' }
+
+        wrap_deferred 'when the option is defined' do
+          include_deferred 'should define component option',
+            'example_option',
+            default: 'value'
+        end
+      end
+
+      describe 'with name: a string' do
+        it 'should return the name of the generated method' do
+          expect(described_class.option(name, **meta_options)).to be name.intern
+        end
+
+        include_deferred 'should check for duplicate options'
+
+        wrap_deferred 'when the option is defined' do
+          include_deferred 'should define component option', 'example_option'
+        end
+      end
+
+      describe 'with name: a symbol' do
+        let(:name) { super().intern }
+
+        it 'should return the name of the generated method' do
+          expect(described_class.option(name, **meta_options)).to be name
+        end
+
+        include_deferred 'should check for duplicate options'
+
+        wrap_deferred 'when the option is defined' do
+          include_deferred 'should define component option', 'example_option'
         end
       end
     end
