+++ spec/librum/components/bulma/label_spec.rb
@@ -1,13 +1,12 @@
 # frozen_string_literal: true
 
 require 'librum/components/bulma/label'
-
-require 'support/deferred/bulma_examples'
-require 'support/deferred/component_examples'
+require 'librum/components/rspec/deferred/bulma_examples'
+require 'librum/components/rspec/deferred/component_examples'
 
 RSpec.describe Librum::Components::Bulma::Label, type: :component do
-  include Spec::Support::Deferred::BulmaExamples
-  include Spec::Support::Deferred::ComponentExamples
+  include Librum::Components::RSpec::Deferred::BulmaExamples
+  include Librum::Components::RSpec::Deferred::ComponentExamples
 
   subject(:component) { described_class.new(**component_options) }
