+++ spec/librum/components/bulma/page/header/brand_spec.rb
@@ -2,13 +2,12 @@
 
 require 'librum/components/bulma/page/header/brand'
 require 'librum/components/literal'
-
-require 'support/deferred/component_examples'
+require 'librum/components/rspec/deferred/component_examples'
 
 RSpec.describe Librum::Components::Bulma::Page::Header::Brand,
   type: :component \
 do
-  include Spec::Support::Deferred::ComponentExamples
+  include Librum::Components::RSpec::Deferred::ComponentExamples
 
   subject(:component) { described_class.new(**component_options) }
