+++ spec/librum/components/bulma/page/header_spec.rb
@@ -2,11 +2,10 @@
 
 require 'librum/components/bulma/page/header'
 require 'librum/components/literal'
-
-require 'support/deferred/component_examples'
+require 'librum/components/rspec/deferred/component_examples'
 
 RSpec.describe Librum::Components::Bulma::Page::Header, type: :component do
-  include Spec::Support::Deferred::ComponentExamples
+  include Librum::Components::RSpec::Deferred::ComponentExamples
 
   subject(:component) { described_class.new(**component_options) }
