+++ spec/librum/components/icon_spec.rb
@@ -2,11 +2,10 @@
 
 require 'librum/components/icon'
 require 'librum/components/icons/font_awesome'
-
-require 'support/deferred/component_examples'
+require 'librum/components/rspec/deferred/component_examples'
 
 RSpec.describe Librum::Components::Icon, type: :component do
-  include Spec::Support::Deferred::ComponentExamples
+  include Librum::Components::RSpec::Deferred::ComponentExamples
 
   subject(:component) { described_class.new(**component_options) }
