+++ spec/librum/components/icons/base_spec.rb
@@ -1,11 +1,10 @@
 # frozen_string_literal: true
 
 require 'librum/components/icons/base'
-
-require 'support/deferred/component_examples'
+require 'librum/components/rspec/deferred/component_examples'
 
 RSpec.describe Librum::Components::Icons::Base do
-  include Spec::Support::Deferred::ComponentExamples
+  include Librum::Components::RSpec::Deferred::ComponentExamples
 
   subject(:component) { described_class.new(**component_options) }
