--- spec/support/deferred.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-module Spec::Support
-  module Deferred; end
-end
